requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-api-core==2.22.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
idna==3.10
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pypdf==5.1.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
urllib3==2.2.3
